# pkg/blackrock_io/spec.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "spec.proto",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/push": {
      "post": {
        "operationId": "SayPush",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ioSuccess"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": " (streaming inputs)",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ioEnvelope"
            }
          }
        ],
        "tags": [
          "Search"
        ]
      }
    },
...
  },
  "definitions": {
...
    "ioEnvelope": {
      "type": "object",
      "properties": {
        "metadata": {
          "$ref": "#/definitions/ioMetadata"
        },
        "payload": {
          "type": "string",
          "format": "byte"
        }
      }
    },
...
    "ioKF": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "value": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "ioKV": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "ioMetadata": {
      "type": "object",
      "properties": {
        "search": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ioKV"
          }
        },
        "count": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ioKV"
          }
        },
        "properties": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ioKV"
          }
        },
        "numeric": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ioKF"
          }
        },
        "created_at_ns": {
          "type": "string",
          "format": "int64"
        },
        "event_type": {
          "type": "string"
        },
        "foreign_id": {
          "type": "string"
        },
        "foreign_type": {
          "type": "string"
        },
        "track": {
          "type": "object",
          "additionalProperties": {
            "type": "integer",
            "format": "int64"
          }
        },
        "id": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
...
    "ioSuccess": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
...
  },
...
}